Back-End/requirements.txt
langchain
langchain_community
python-dotenv
langchain_text_splitters
langchain_huggingface
sentence_transformers
chromadb
faiss-cpu
langchain-openai
langchain_ollama
langchain_groq
